vineyard-backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
